pyproject.toml
[project]
name = "pwc-genai-task"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docling>=2.68.0",
    "dropbox>=11.36.0",
    "google-api-python-client>=2.0.0",
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.8",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.7",
    "langchain-text-splitters>=1.1.0",
    "langgraph>=1.0.6",
    "openai>=2.15.0",
    "python-dotenv>=1.2.1",
    "requests>=2.31.0",
    "streamlit>=1.28.0",
    "tqdm>=4.67.1",
]
